yang: Allow writes to interface name and type
We are gradually introducing configuration support for parts of
ietf-interfaces tree.
Let's start with a constraint for names and types of the interfaces.
We don't want users to introduce new interfaces or delete the existing
ones.

We don't want users to reconfigure the lo (loopback) link, therefore it is
not listed in the list of allowed interfaces. You can't use that link in
running or startup datastores. Operational datastore is, however, not
validated, hence the loopback can appear there.

Change-Id: I8b2e030842874ff1c2baee72a270e2f680d933fe