system: Hardcode network interface types
This is a follow-up to the previous commit which reverted an attempted
auto-discovery of interface type.
The reverted code was not reliable. The set of links is not expected to
change in the system, hence I think we can initialize it in main and
pass to IETFInterfaces class.

Unfortunately, the list of interfaces and types is duplicated. It
already appears in the YANG model (see the czechlight-network YANG model
which introduces the link name-type constraint).

Change-Id: I9ad31a7d2ab5aa690c203f7adda65f26ae2e79a6