system: Configurable IP addresses for interfaces
We alter deviations in czechlight-network YANG model to allow
configuring IPv4 and IPv6 addresses for a specific interface.

The ietf-ip:ipv{4,6} presence containers (and the 'enabled' leaf in
those containers) also control whether the protocol is enabled. If the
protocol is disabled for a link but some addresses are still present,
we must not configure any address for such link.

Also temporarily disable DHCP client.

Change-Id: Ic2d1198122707a5606482fec2a05a44726c812b3