system: with ietf-interfaces
When one removes ietf-ip:ipv4 (ietf-ip:ipv6) containers, the
IPv4 (IPv6) connectivity is disabled. We can foresee some
accidental connectivity losses here so let's make it a little bit harder
for users to achieve that.

We require that at least one of ietf-ip:ipv4 and ietf-ip:ipv6 presence
containers is present. If not, then ietf-interfaces:enabled leaf must be
set to false (this disables the interface) or the interface is enslaved
(not yet implemented as the bridge functionality is not yet
implemented).

Of course, one can delete all addresses, turn off the autoconfiguration
on a not-enslaved link and the connection will be lost anyway.

Change-Id: Ieef254644ee4de107675b8e67d86b2ed7dfec3ab