tests: Fixes in interface-sudo tests
We have seen several failures in the CI system regarding the
interfaces-sudo test. Sometimes the expected values were not correct. It
looked like we don't give system enough time to process the bridge
configuration changes.

When I increased the timeout to 3 seconds, another issue appeared. The
bridge obtained an IPv6 LL address which was not expected by the test.
So let's fix that too.
Hopefully, this fix combined with the increased wait time will resolve
the CI issues.

Change-Id: Ida1d4ab6d7958c5e0e187299e765a4f6e25bc3af