system: Network autoconfiguration
Add a czechlight-network:dhcp-client leaf into
ietf-interfaces:interfaces/interface/ietf-ip:ipv4 container which
indicates whether the implementation starts an DHCP client on IPv4.

For IPv6, we use
ietf-interfaces:interfaces/interface/ietf-ip:ipv6/autoconf container.
The value of autoconf/create-global-addresses leaf is mapped to
networkd's IPv6AcceptRA config value [1].

The 'DHCP=' option in systemd.network [2] allows multiple values
managing DHCPv4 and DHCPv6. We chose to ignore DHCPv6 because it will
(or won't) be triggered by router advertisement (RA) *regardless* of
this parameter [2]. It is sufficient to have RA on (which is managed by
the ipv6/autoconf leaf). In case RA is on but no routers are found,
systemd-network starts the DHCP client anyway [1]. If RA is off, then
no IPv6 autoconfiguration happens.

[1] https://systemd.network/systemd.network.html#IPv6AcceptRA=
[2] https://systemd.network/systemd.network.html#DHCP=

Change-Id: Ia885ae644d368987b6101f828bddfd0fc4e969dd