system: Move lldp-systemd-networkd-sysrepo to velia
We agreed that maintaining standalone lldp-systemd-networkd-sysrepo
project is excessive. For example, many compilation units in the
project are copied from velia. Also, velia manages a lot of
system (and network) related things so why should the access to LLDP
neighbours via sysrepo be a standalone project?

This commit copies lldp-systemd-networkd-sysrepo project into velia.
The source repository commit is
39bd89ff7e57cd7bf4ebfdab5fce41fb4fdedcf0.

Only necessary stuff was changed (some naming and compilation issues).
We *might* be changing this code soon if our patch for LLDP neighbours
extraction in JSON format[1] is merged into systemd.

Of course, this also needs a br2-external update because this commit
effectively deprecates lldp-systemd-networkd-sysrepo project and it
should not be used in br2-external anymore.

[1] https://github.com/systemd/systemd/pull/20333

Change-Id: Ide2ed4e5eb8f1c3b10f0e2af7820f83c04cb81e8