utils: refactor utils::getSubtree libyang wrapper
In the change for the commit to follow Vasek suggested that
the utils::getSubtree helper could be renamed to getUniqueSubtree and
the behaviour could be changed to return std::optional<>. When no
matching node is found the function returns std::nullopt. When multiple
matching nodes are found the function throws.

This code in the next commit could benefit a little from this change.

Change-Id: I8ded7695dc383e6654f8efa44b6a5442ef904485