Change date-and-time format
According to RFC 6991[1], the canonical format for date-and-time is
local time. That means that when printing the canonical value, the
string is adjusted to the timezone of the machine. This is undesirable
for testing and overall (we want everything to be UTC). Specifying "Z"
means that the timezone is known and is UTC. On the other hand,
specifying "-00:00" means that the timezone is unknown and no timezone
correction should be done.

[1]: https://datatracker.ietf.org/doc/html/rfc6991

Change-Id: Idf956624ca56431f232a20c33e55f69311e2e51c