tests: Attempt to fix race in rauc test
CI tests with asan/ubsan sometimes fail in system_rauc test. [1]

The test log looks like this:

```
 /home/ci/src/cesnet-gerrit-public/CzechLight/velia/tests/system_rauc.cpp:8:
 TEST CASE:  Fetch RAUC data over DBus
   Installation OK

 /home/ci/src/cesnet-gerrit-public/CzechLight/velia/tests/system_rauc.cpp:142: FATAL ERROR: REQUIRE( rauc->operation() == "installing" ) is NOT correct!
   values: REQUIRE( idle == installing )

 (...)
```

By looking at the code, it seems that the contents of the REQUIRE macro
 REQUIRE(rauc->operation() == "installing");
is executed before the RAUC installation actually starts (it is executed
from another thread) and therefore before it sets current DBus property
"Operation" to value "installing".
Let's try to fix that by waiting for the installation to actually start.

[1] https://object-store.cloud.muni.cz/swift/v1/ci-logs-public/70/5270/2/check/f34-clang-asan-ubsan/4467b92/job-output.txt

Change-Id: I8e6e217660f3b87045a5538949a84139573a005a