Require sdbus-cpp version 1.0.0
We have seen a strange issue in some of the CI tests [1].
Our D-Bus client obtained a DBus signal that it should not have
obtained (UnitNew for
/org/freedesktop/systemd1/unit/sys_2ddevices_2dvirtual_2dnet_2dczechlight0_2edevice)
which is a unit for a new interface added in a parallel test to a real
system, not our testing systemd server.
The signal matched all the interfaces and object paths,
but it was issued from another bus (real systemd bus rather than our
testing bus which is just mimicking the behaviour of real systemd
server).
I think we are hitting bug [2] from sdbus-cpp. The bug is fixed so let's
see.

[1] https://object-store.cloud.muni.cz/swift/v1/ci-logs-public/89/5289/3/check/f34-gcc-cover/31c8db3/job-output.txt
[2] https://github.com/Kistler-Group/sdbus-cpp/issues/145

Depends-on: https://cesnet-gerrit-czechlight/c/CzechLight/dependencies/+/5321
Depends-on: https://cesnet-gerrit-public/c/CzechLight/dependencies/+/5321
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/5321
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/cla-sysrepo/+/5309
Change-Id: If44ada7bf6eae5c73b87daf0f5171160f95d38f5