tests: easier mocking of notification handling
I think it's much, much easier to use features of the mocking framework
we have *and* the modern C++ bindings. While this removes some
features (time tracking, could be easily added when needed), it also
cleans up bits which were never really used before, such as notification
replay and explicit start/stop operations.

The good thing is that we can now delegate partial matching to the
trompeloeil and its WITH macro, for example.

Change-Id: Id7e4b73575de4a1b1fbe38d4a0316ae75bd677dc