RAUC: change slot status reporting
I would like to make it possible to change which one of the healthy
slots gets booted next time, but for that, we must keep track of which
slots are healthy. I think that in this context exporting them as
booleans makes more sense, and it will make these property data types a
bit less complex.

Change-Id: I8ebd9c28a4c5b5717e9d61ae5693ef506d72e905