health: Report systemd unit states with alarms
We have been developing sysrepo-ietf-alarms project which implements
important parts of the ietf-alarms YANG model.
Let's use ietf-alarms for reporting states of systemd units. If the unit
is not healthy, we issue or clear alarm (by calling the RPC located at
/sysrepo-ietf-alarms:create-or-update-alarm).

One of the next commits will rewrite the LED manager to use alarms.
The original health manager will be scraped after that.

Change-Id: I08eb07f3d02836b77ca2f43b2cca49469264c97a