health: sysrepo daemon for LED controlling based on alarm count
This patch introduces a daemon that listens for changes in ietf-alarms
operational datastore and whenever alarm count changes it triggers a
callback. We write this code with the intention to use it for
controlling LED colour but it is generic so one can plug another stuff
into that.

For now we only react on change in the alarm count and we light only
green and red colour. This is to be changed soon(TM) as we want it to
light up different colours based on the alarm list.
If there are critical alarms we should light the red colour. If there are
less important alarms only we should light the orange colour to warn the
user. For no alarms we consider green light to be best.
However, this requires grouping alarms based on the severity. We will
implement alarm-summary feature of the ietf-alarms model in
sysrepo-ietf-alarms first so we don't have to implement the logic for
grouping alarms by severity twice.

Change-Id: I5958975f246f96762389aae27b1974d5eea4093d