health: control LEDs via sysrepo and ietf-alarms
This patch introduces a sysrepo subscription that listens for changes
in ietf-alarms' operational datastore and whenever the alarm count
changes it triggers a callback. We write this code with the intention
to use it for controlling LED colour but it is generic so one can plug
another stuff that somehow reacts to the number of alarms.

For now we only react on changes in the alarms list. We consider only
uncleared alarms. We set the output state to OK or ERROR based on the
count of alarms (if 0 alarms then OK else ERROR) and we let the output
signal to handle the state change somehow (e.g. light up different LED).

Change-Id: I5958975f246f96762389aae27b1974d5eea4093d