health: set LED color based on alarm-summary
This patch changes the strategy of setting of alarm reporting by LEDs.
Before this patch we just set green or red based on the information
whether there are *any* active (not-cleared) alarms.

However, ietf-alarms YANG model supports alarm-summary feature which
reports the number of alarms grouped by severity. We now support that in
sysrepo-ietf-alarms so let's read the data in velia.

We also turn on the option of setting the LED color to orange. It is
used if the most severe active alarm is of severity WARNING and lower.

Depends-on: https://gerrit.cesnet.cz/c/CzechLight/sysrepo-ietf-alarms/+/5878
Change-Id: I3eab7aaeebff891df21392e7be29e47b463f45f7