system: remove fmtlib version check
We had this ifndef in the code that was introduced to allow building
with both fmt<8 and fmt==8.

It was there due to change in fmt==8 because fmt::format now requires
the format string to either be known at compile-time or to be wrapped
in fmt::runtime [1,2].

Now that we are all (incl. CI) on fmt>=8 we can safely remove that.

[1] https://github.com/fmtlib/fmt/releases/tag/8.0.0
[2] https://fmt.dev/8.1.1/api.html#core-api

Change-Id: I70963ff3e44dabb91bf7fff78f18bec83351be73