tests: run sudo tests in container rather than under sudo with cleanup
There is a test suite where we manipulate with network stack and we
check that velia correctly captures such changes and propagates them to
sysrepo. However, to manipulate with network stack one must have
privileges for that and also such manipulations *may* destroy some
network settings on the host computer. That's why these tests were not
automatically run but they had to be explicitly enabled in
configuration.

jkt proposed a better way: Run this in a network namespace using
unshare[1]. This way we can run these tests always and not worry about
destroying any network settings.

There is one caveat though. When sysrepo is compiled with
`-DSYSREPO_GROUP=<group>` then the files it creates are owned by
<user>:<group>. But `unshare --map-root-user` only maps the <user> to
root and not the <group> (unless it is your main group).
Hence we explicitly call unshare with `--map-group=<group>` when we
detect that sysrepo was indeed compiled with `-DSYSREPO_GROUP`. This way
we ensure to have correct permissions inside the namespace to open
sysrepo files which were created before the namespace.

[1] https://www.man7.org/linux/man-pages/man1/unshare.1.html

Change-Id: Ifb19c1b49335a3ae2580cb5c8c34aeda8f8b7b52