tests: do not silently ignore a mount failure
I have a weird system, and my `mount` assert-crashes in the namespaced
environment. That's of course a bug in my system, but because the test
suite uses a pattern of `sh -c 'foo; bar'`, the failure of `foo` is
masked.

Fix that by asking the shell to die early, and while we're at that,
let's be a bit more verbose and print out the commands as they are
executed.

Change-Id: I16100b43b6221f6f0e0032cdd8ba6cdcf5fe7f09
Fixes: fd5ec7b tests: run sudo tests in namespace rather than under sudo with cleanup