tests: merge tests for ietf-interfaces module
We had two tests for ietf-interfaces module. One of them required sudo
to be able to manipulate with network interfaces settings and we didn't
like having it in our default suite so it had to be explicitly enabled.

Since commit fd5ec7b700dcebcfc43ad5d5cc265aceb3c772d8 this no longer
requires sudo because it runs in a namespace. There is no reason to
keep those tests separately so this commit merges these two tests.

Change-Id: If2ce7074c863e64c1c4d4be9ed0dd27dd5cd0c48