tests: split ietf-hardware further
The ietf-hardware test now contains testing of both the IetfHardware with
DataReaders (which are responsible for correct reading of the hardware
data from hwmon), and the Sysrepo class which is responsible for
publishing the data processed by IetfHardware into sysrepo.

We are going to make changes to how Sysrepo class behaves which will
require slightly bit different testing approach. Testing both of those
classes in one test case might complicate the test, so splitting it
might be a good idea that would not hurt.

Change-Id: I3083c66b098458f284ff23aed4daf33e4ad2c0ee