hardware: add thresholds watchers
This patch series will start publishing alarms on hardware sensors. The
sensor values should have some thresholds that trigger the alarms.
However, simply comparing the sensor value to threshold is probably not
enough because we don't want keep creating and deleting alarms when the
value is dancing around the threshold. There should be some code that
triggers the alarm based on the history of the values.

Fortunately, such code already exists in our cla-sysrepo project. This
commit copies that from cla-sysrepo (commit
890d75d99a3ecaa6b8fc9d9a269ad70f57f69e49) into velia-hardware.

Change-Id: I563006b093b190a16c87558c92956ef4a49382b2