hardware: support component/state container in ietf-hardware
Our PSU can be plugged or unplugged when the system is running. When
this happens whole PSU subtree is not returned to sysrepo from
ietf-hardware.

We decided to change that and use state container with oper-state leaf
in ietf-hardware[1] that reports the operational status of a component.
In this commit, we set every oper-state leaf to enabled but in the next
commit we implement setting oper-state of our PSU to enabled/disabled
depending on whether it is/isn't currently present.

[1] https://www.rfc-editor.org/rfc/rfc8348

Change-Id: I27d8e3d2bf66592fdcfdaaca992a4b5b5f10cb73