hardware: set oper-state to disabled if PSU is not present
Our PSU can be plugged or unplugged when the system is running. When
this happens whole PSU subtree is not returned to sysrepo from
ietf-hardware.

This patch changes that by returning the PSU component but with no
sensors and oper-state [1] set to disabled.
We did not discover any other possibility of how to report a
"not-present".

Originally I thought that a clever rewrite of ietf-hardware
abstraction that somehow polls the current "state" of the component
would be a good idea. But the code reporting PSU component works
differently than the rest of the sensors and we only want to support
disabled/enabled for this component. Therefore I only simply edited the
PSU code.

[1] https://www.rfc-editor.org/rfc/rfc8348

Change-Id: I6a8a992105d42df7491b1c10614ead680836de65