hardware: data readers can set up sensor thresholds
We are monitoring some hardware sensors. Some sensor values might be
considered dangerous (e.g. high temperatures). This commit introduces a
way of storing thresholds for data readers and validating measured
sensor values against those thresholds.

This is implemented in a way that each data reader must also implement a
method called thresholds() that returns a mapping of sensor XPath to
a Threshold object. These will be collected in the future by
IETFHardware and used to check obtained sensor values in order to raise
alarms.

Change-Id: I2d3854603f2ec04613906290633151f4d7b034f1