hardware: store and process threshold watchers
We register the thresholds for sensors that are reported by individual
data readers and we match them against sensor values.

This code does not yet set up any alarms but IETFHardware::process now
returns not only the ietf-hardware YANG data but also important changes
in values that can be used to raise alarms by the caller.

Change-Id: I2660a411be517ba104f5c05bba1c6e97053569ad