hardware: add sensor related alarms
We add two possible alarms our hardware processing can trigger:

 - Sensor reading is above (or below) specified thresholds, and
 - sensor is missing.

We fill the alarm-inventory with all possible alarms-components
pairs.

At first glance there is an unrelated change to when the background
polling thread is spawned.
But we need first to fill the alarm-inventory and only after that the
first alarms can be pushed into sysrepo. From my understanding of
ietf-alarms [1] alarms must be listed before they are added.
That is why the spawn of the background polling thread is delayed.

[1] https://www.rfc-editor.org/rfc/rfc8632

Change-Id: I8a83425e5a559029b29d7e2c39fbf934dedd94f0