tests: increase wait time in czechlight-system:leds test
I have seen the following errors in test-sysrepo_system-leds several
times.

  2023-05-11 12:26:19.979398 | w | /home/ci/src/cesnet-gerrit-public/CzechLight/velia/tests/test_sysrepo_helpers.h:21: FATAL ERROR: REQUIRE( data ) is NOT correct!
  2023-05-11 12:26:19.979402 | w |   values: REQUIRE( {?} )

This means that the data in sysrepo were not already there when we asked
for them. This might be caused by the slow thread startup. Let's make
sure there is enough time and sleep everywhere where the sleep is
needed. It does not matter in tests.

Change-Id: I850c056d8a29d4b6bf55614d6f206cf0b45c06e9