hardware: allow no value in threshold watcher
This patch allows updating threshold watcher with no value. This is
beneficial for us in hardware monitoring. Upon disconnecting the sensor,
its value is no longer present and no longer updated. We therefore have
to store somewhere that the value is not there.
One of the easy options is to set Watcher's state to "NoValue" which
tells us that the sensor is not there.

When the sensor is connected again we don't care about previous history
and set the state as if there was no history (see tests).

Change-Id: I8d7a2b31b13ded82c1f9d7e83ea48db32b8cf347