hardware: invoke sensor threshold crossing alarms
Enable reporting alarms about crossing threshold values. We use alarms
defined in previous commits. We use same alarm for warning and critical
crossings of thresholds, only the severities change.

The only caveat here was a limitation of our RPC-based alarm API from
sysrepo-ietf-alarmsd. The problem happens when the value changes
straight from low to high (or the other direction). However, the RPC
can only create/modify one alarm per one invocation so we can't clear
the low alarm and set the high alarm by one sysrepo edit (to make it
appear as both changes happened at one time). Instead we first set the
new alarm and then clear the old alarm so that the resource is in
"alarmable" state for the whole duration.
Perhaps we should change the alarms RPC to accept a list of changes
instead of a single change.

Change-Id: I28bafcc498e5887b70584d86aa1a0e6ac25f03ea