log: ensure we can control verbosity of everything
The code uses three loggers:

- `system`/`firewall`/`health`/`hardware` for the application-specific
  part of each daemon
- `main` for exception handling and a wrapper around calls to
  sysrepo/libyang
- `sysrepo` for the C library

I think it makes sense to keep that separation between
sysrepo-the-C-library and our-wrapper-around-calling-sysrepo-functions.
For a lack of a better name, this is called `main`. That's no change,
all this commit is doing is making sure that the actual log verbosity
setting is possible in all daemons. Without this patch, I don't know
what causes some failure in velia-health because I don't know what call
to what XPath yields an error from libyang.

Change-Id: Ifa5dda96267eae6ad6afcfce6c61ada3621cda8a