Fix build failure with GCC 13.2
After a system update, this stopped building:

 velia/src/system/Authentication.cpp: In function ‘auto velia::system::impl::file_open(const char*, const char*)’:
 velia/src/system/Authentication.cpp:58:65: error: ignoring attributes on template argument ‘int (*)(FILE*)’ [-Werror=ignored-attributes]
    58 |     auto res = std::unique_ptr<std::FILE, decltype(&std::fclose)>(std::fopen(filename, mode), std::fclose);
       |                                                                 ^

The root cause is that fclose's argument has
__attribute__((nonnull(1))), and that attribute triggers that warning.

It turns out that this pattern is actually not strictly speaking correct
because one is not allowed to take the address of a standard function
like this. Also, declaring the unique_ptr with a function-pointer
destructor is a wee bit less effective because this type would offer
runtime customization of its behavior, whereas we do not really need
to consult this function pointer at runtime.

Fortunately, as of C++20 it's OK to use a decltype(lambda) as the
second template type to unique_ptr (and then it's OK to rely on its
implicit "construction" for the actual destructor), so the fix is pretty
short. Yay!

Change-Id: Ibd0397ad3446c0fe95de297f36d11637eb74d96c
Link: https://stackoverflow.com/questions/76867698/what-does-ignoring-attributes-on-template-argument-mean-in-this-context
Link: https://stackoverflow.com/a/76867913/2245623