hardware: set sensor's oper-status when adding sensor value
It might happen that the sensor reading is obviously faulty (e.g. the
value is way out of range). So we want to set the sensor's oper-status
to nonoperational. Currently, we were setting the oper-status to "ok"
when the data reader was initialized. Now, it is set upon each reading
of the sensor value.

Change-Id: I3f0968e8a7bbeaa8b4e9418a3b817692856eaf16