hardware: addSensorValue overload for strings and ints
We will want to check the actual numeric value passed to
addSensorValue function in the next commit to detect
underflowed/overflowed values.

The function addSensorValue, which writes the sensor-value into the
tree, now accepts strings only. Most of our sensor values are integer so
we convert them to strings.
This commit adds new addSensorValue overload for integers to avoid
converting ints to strings. Now, we can easily check the value passed to
detect overloads.

However, this commit does not check if the passed string value is really
a valid numeric value.

Change-Id: I8c32f063fa92c7f45ecb1a21364276f4be780868