hardware: addSensorValue overload for strings and ints
We will want to check the actual numeric value passed to
addSensorValue function
I think we don't want to convert the sensor int value to a string and
then back to int to check the value and then again to string... This
might be a little better.
However, we will do that for values that we fetch as strings anyway.

If the value passed as a string is not numeric, then log an error.

Change-Id: I8c32f063fa92c7f45ecb1a21364276f4be780868