hardware: Disallow writing out-of-range values into sensor-value leaf
One of our boxes started reporting nonsense value of one of the sensors.
This value is way beyond range for sensor-value and it led to crashes
because libyang refused to write such value into the sensor-value leaf
and throwed with LY_EVALID.

Let's hotfix that for now to set value to zero and sensor oper-status to
nonoperational.

However, the problem of setting the correct value is more complicated
because the valid range for the int32 value in YANG is one thing but the
descriptions gives even more constraints depending on sensor-value-type.

Change-Id: I9a74991da58239e0f376f2ec3e1e86f88eab9c57