hardware: add new sensors to alarm-inventory
After previous commits we saw failures on live boxes that nonoperational
alarm is not registered in alarm-inventory for some components.
And indeed. We fill alarm-inventory right after the Hardware classes
are constructed. However, when the pluggable components are not present
at that time, then its thresholds would not get passed to IETFHardware.
And we read active sensors based on the keys of sensor XPath to
Thresholds mapping.

The simple solution was to load all the active sensors and their
thresholds with every HW state poll. Then, we'd sync the
alarm-inventory with the current state of the active sensors. However,
this could not work. If the sensor is unplugged, we'd delete all
alarm-inventory entries for this sensor but then we can't publish the
missing sensor alarm.

Therefore, let's do it like in health tracking alarms for systemd
services: Publish new alarm-inventory entries but do not delete them.

Change-Id: I8a9b85e3b11f3ea4d66a48241d39b808a8b5dffa