hardware: make FspYhPdu hot-pluggable
We have seen pmbus failures on our boxes. The PDU hwmon is not present.
This commit makes the PDU handling similar to PSUs, i.e., it will
consider PDU as hot-pluggable.

This is not a fix of the problem but rather a workaround so that
velia-hardware does not fail because the expected HWMon directory is not
present.

Change-Id: I76338dea8c96a3d2d281a206ba0b254749fdc5d4