tests: reduce excessive whitespace
I know that clang-format is happy to align the macro trailing
backslashes, but enough is enough. Also, having a source file "correctly
formatted" by a tool is not a good thing if the text starts at column
105.

Also, clang changed the AlignEscapedNewlinesLeft to an enum
AlignEscapedNewlines, and IMHO this new option is what makes most sense.
Still, it would try to rewrap some stuff for me, but I think this is a
good step.

Change-Id: Id03e489ada2b51ba4698ce263f0a915cebf27ef6