tests: verify that called alarms are registered in alarm-inventory
I was writing a code that raises alarms but forgot to register the alarm
in the alarm-inventory. When this happens the tests pass. However, on
real devices the RPC call actually ends with a failure. The alarm must
be registered in the alarm-inventory.

Let's make sure that tests also require raised alarms to be registered
in the alarm-inventory. Luckily, we can use some fancy trompeloeil
features for that.

Change-Id: I1e0b48c4f417ee51e2162facd62ae7024101a8dd