tests: use alarm expectations in systemd-unit tests
I will be rewriting the class that is tested by this code soon. So far,
we only check how the datastore looks after performing some operations.
However, I would like to be sure that I see all the changes made to a
datastore and verify that all RPCs are indeed called and the
alarm-inventory is filled with the correct data.

Change-Id: Iffcea3208c299e926295a1913503ede1feb51bd4