tests: mock class for testing alarms
I want to create a class for testing alarms. This means, we need a
simple mocked alarm server that listens for alarm RPCs and watches for
alarm inventory changes in the datastore. Fortunately, we have both of
these classes already, so it is just a matter of combining these two
watchers and an AlarmInventory class introduces in (b0923ecc "tests:
verify that called alarms are registered in alarm-inventory").

Change-Id: I914d6ee6d941ec942288df3093a1f95ffbe30591